pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "root3st"
version = "1.0.0"
description = "A comprehensive OSINT reconnaissance tool for IPs, domains, emails, usernames, and social media profiles"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Root3st Contributors"},
]
keywords = ["osint", "reconnaissance", "security", "intelligence"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "requests>=2.31",
    "dnspython>=2.4",
    "python-whois>=0.9",
    "beautifulsoup4>=4.12",
    "aiohttp>=3.9",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.1",
    "responses>=0.24",
]

[project.scripts]
root3st = "root3st.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["root3st*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
